pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sage"
version = "0.1.0"
description = "CLI coding assistant"
dependencies = [
    "google-generativeai>=0.3.0"
]

[project.scripts]
sage = "sage.cli:main"
